+++ au3/libraries/lib-builtin-effects/CMakeLists.txt
@@ -63,8 +63,6 @@ set( SOURCES
    RepeatBase.h
    ReverbBase.cpp
    ReverbBase.h
-   Reverse.cpp
-   Reverse.h
    SBSMSBase.cpp
    SBSMSBase.h
    ScienFilterBase.cpp
